Issue-102: track AfterStageExecution on a per-method basis
This is elemantary to have both repeatable and non-repeatable methods in the same Stage class.

Furthermore since the Stage state has now grown immensly in complexity (it is now a real object) move the stage state two its own class and disregard the strong ties (especially Scenario Executors' write access to a field) for now.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>